requirements.txt
fastapi
uvicorn
python-dotenv
streamlit-authenticator==0.3.3
